\name{cups}
\docType{data}
\alias{cups}
\title{Dataset: RomanCups}
\description{This data set consists of the compositions of 11 oxides in 47 Roman cups found at an archaeological site in eastern England. Compositions are expressed as percentages.}
\usage{data(cups)}
\format{Data frame containing the 47 x 11 matrix.}
\source{Baxter MJ, Beardah CC, Cool HEM and Jackson CM (2005) Compositional data analysis of some alkaline glasses. Mathematical Geology 37: 183-196.}
\keyword{datasets}
