\name{fish}
\docType{data}
\alias{fish}
\title{Dataset: FishMorphology}
\description{This data set consists of the sex, habitat, mass and then 26 morphometric measurements on 75 fish (Arctic charr)}
\usage{data(fish)}
\format{Data frame containing the 75 x 29 matrix. Column 1 contains sex (1=female, 2=male). Column 2 contains habitat (1=litoral, 2=pelagic). Column 3 contains the mass in grams. Columns 4 to 29 contain the 26 morphometric measurements.}
\source{Greenacre, M and Primicerio, R (2010) Multivariate Analysis of Ecological Data. BBVA Foundation, Bilbao. Free download at www.multivariatestatistics.org}
\keyword{datasets}
