\name{Tellus}
\docType{data}
\alias{Tellus}
\title{Dataset: Tellus}
\description{This data set consists of the molar values of 46 minerals in 6643 samples of topsoil in Northern Island. In addition, 11 environmental variables are provided.}
\usage{data("Tellus")}
\format{Data frame containing the 6643 x 57 matrix. First 11 variables are environmental, variables 12 to 57 are the molar values.}
\source{Personal communication from Eric Grunsky.}
\keyword{datasets}
