\name{time}
\docType{data}
\alias{time}
\title{Dataset: TimeBudget}
\description{This data set consists of the average percentage breakdown of time use into six categories, for 16 countries, split by males and females.}
\usage{data(time)}
\format{Data matrix containing the 32 x 6 matrix. Row samples are labelled by the two-character country code and m (male) or f (female).}
\source{Greenacre M., Compositional Data Analysis in Practice, Chapman & Hall / CRC, 2018.}
\keyword{datasets}
