\name{veg}
\docType{data}
\alias{veg}
\title{Dataset: Vegetables}
\description{This data set consists of the protein, carbohydrate and fat compositions of 10 different vegetables. Compositions are expressed as percentages.}
\usage{data("veg")}
\format{Data frame containing the 10 x 3 matrix.}
\source{US Department of Agriculture, https://ndb.nal.usda.gov/ndb/nutrients/index}
\keyword{datasets}
