\name{fish}
\docType{data}
\alias{fish}
\title{Dataset: FishMorphology}
\description{This data set consists of the sex, habitat, mass and then 26 morphometric measurements on 75 fish (Arctic charr)}
\usage{data("fish")}
\format{Data frame containing the 75 x 29 matrix.}
\source{Greenacre, M and Primicerio, R (2010) Multivariate Analysis of Ecological Data. BBVA Foundation, Bilbao. Free download at www.multivariatestatistics.org}
\keyword{datasets}
